requirements.txt
black==24.8.0
fpdf==1.7.2
matplotlib==3.9.2
pytest==8.3.3
ruff==0.6.5
jupyter==1.1.1
nbval==0.11.0
polars-lts-cpu==1.7.1